Fix stack overflow during disconnect
Disconnect call has a stack frame depth of 23 till del_key in driver,
and with default stack size of 2048 (non-openthread) this causes a stack
overflow, so, increase the SHELL stack size for now, but this then moves
the stack overflow to syswork_q, so, increase the stack size for
SYSTEM_WORK_QUEUE as well.

In future we can add a max stack frame depth and optimize esp. the
driver_if in wpa_supplican to the driver interface.